requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
cached-property==1.3.1
celery==3.1.18
certifi==2015.11.20
Django==1.9.5
django-celery==3.1.17
django-redis==4.4.1
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
docker==2.4.2
docker-compose==1.14.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
gunicorn==19.3.0
imutils==0.4.3
ipaddress==1.0.18
Jinja2==2.9.6
jsonschema==2.6.0
kombu==3.0.34
protobuf==3.3.0
psycopg2==2.7.3
PyJWT==1.5.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==15.2.0
redis==2.10.5
requests==2.11.1
six==1.10.0
stevedore==1.25.0
subprocess32==3.2.7
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
websocket-client==0.44.0
Werkzeug==0.12.2
uwsgi==2.0.15
elasticsearch==1.9.0
elasticsearch-dsl==0.0.11
pymongo==2.7.1
six==1.10.0
mongoengine==0.9.0
django-rest-framework-mongoengine